renamed temp to thermal
