+++ src/jcarbon-proto/src/main/java/jcarbon/BUILD
@@ -7,7 +7,7 @@ java_library(
         "//src/jcarbon-proto/src/main/java/jcarbon/linux/jiffies",
         "//src/jcarbon-proto/src/main/java/jcarbon/linux/powercap",
         "//src/jcarbon-proto/src/main/java/jcarbon/linux/freq",
-        "//src/jcarbon-proto/src/main/java/jcarbon/linux/temp",
+        "//src/jcarbon-proto/src/main/java/jcarbon/linux/thermal",
         "//src/jcarbon-proto/src/main/java/jcarbon/emissions",
         "//src/jcarbon-proto/src/main/java/jcarbon/rapl",
         "//src/jcarbon-proto/src/main/java/jcarbon/util",
@@ -17,7 +17,7 @@ java_library(
         "//src/jcarbon-proto/src/main/java/jcarbon/linux",
         "//src/jcarbon-proto/src/main/java/jcarbon/linux/jiffies",
         "//src/jcarbon-proto/src/main/java/jcarbon/linux/powercap",
-        "//src/jcarbon-proto/src/main/java/jcarbon/linux/temp",
+        "//src/jcarbon-proto/src/main/java/jcarbon/linux/thermal",
         "//src/jcarbon-proto/src/main/java/jcarbon/emissions",
         "//src/jcarbon-proto/src/main/java/jcarbon/rapl",
         "//src/jcarbon-proto/src/main/java/jcarbon/util",
