+++ src/jcarbon-proto/src/main/java/jcarbon/linux/thermal/BUILD
@@ -1,5 +1,5 @@
 java_library(
-    name = "temp",
+    name = "thermal",
     visibility = ["//visibility:public"],
     srcs = glob(["*.java"]),
     deps = [
