+++ src/jcarbon-proto/src/main/java/jcarbon/linux/thermal/TemperatureMonitor.java
@@ -1,4 +1,4 @@
-package jcarbon.linux.temp;
+package jcarbon.linux.thermal;
 
 import static jcarbon.util.LoggerUtil.getLogger;
